total_requirements.txt
# Copyright (c) 2023 Mitsubishi Electric Research Laboratories (MERL)
#
# SPDX-License-Identifier: MIT
huggingface-hub==0.17.0
peft==0.10.0
pytorch_metric_learning
pytorch_warmup
tensorboard
accelerate

charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
einops==0.6.0
filelock==3.10.7
fonttools==4.39.3
ftfy==6.1.1
# huggingface-hub==0.13.3
idna==3.4
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
nltk==3.8.1
numpy==1.21.5
opencv-contrib-python-headless
# opencv-python==4.7.0.72
packaging==23.0
pandas==1.4.4
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
tokenizers==0.19
# tokenizers==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.41.0
# transformers==4.26.0
typing_extensions==4.5.0
urllib3==1.26.15
vit-pytorch==1.2.0
wcwidth==0.2.6
zipp==3.15.0
tiktoken
scipy